requirements.txt
streamlit>=1.32.0
anthropic>=0.40.0
python-docx>=1.1.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
pyperclip>=1.8.2